Move remaining vendor returns to vendor_bp
