+++ API_DETAILS.txt
@@ -112,49 +112,48 @@ POST /vendor/payout/setup
 
 Shop
 ----
-POST /vendor/create-shop
+POST /vendor/shop
     - Auth + vendor role.
     - Body: shop_name, shop_type, optional description, delivers, appointment_only, is_open, category_tags, logo_url, featured, verified.
     - Creates shop for vendor.
     Example request:
-        POST /api/v1/vendor/create-shop
+        POST /api/v1/vendor/shop
         {"shop_name": "My Shop", "shop_type": "grocery"}
     Example response:
         {"status": "success", "message": "Shop created"}
 
-POST /shop/edit
+POST /vendor/shop/edit
     - Auth + vendor role.
     - Body fields to update existing shop owned by vendor.
     Example request:
-        POST /api/v1/shop/edit
+        POST /api/v1/vendor/shop/edit
         {"description": "Open 24/7"}
     Example response:
         {"status": "success", "message": "Shop updated"}
 
-GET /shop/my
-    - Auth + vendor or admin role.
-    - Vendor receives its shop data. Admin receives all shops.
+GET /vendor/shop/my
+    - Auth + vendor role. Returns vendor's shop data.
     Example request:
-        GET /api/v1/shop/my
+        GET /api/v1/vendor/shop/my
     Example response:
         {"status": "success", "data": {...}}
 
-POST /shop/update-hours
+POST /vendor/shop/hours
     - Auth + vendor role.
     - Body: {"weekly_hours": [{"day_of_week": int, "open_time": "HH:MM", "close_time": "HH:MM"} ...]}
     - Replaces stored hours for vendor's shop.
     Example request:
-        POST /api/v1/shop/update-hours
+        POST /api/v1/vendor/shop/hours
         {"weekly_hours": [{"day_of_week": 1, "open_time": "09:00", "close_time": "18:00"}]}
     Example response:
         {"status": "success", "message": "Shop hours updated"}
 
-POST /vendor/shop/toggle_status
+POST /vendor/shop/toggle-status
     - Auth + vendor role.
     - Body: {"is_open": true|false}
     - Marks shop open or closed and logs action.
     Example request:
-        POST /api/v1/vendor/shop/toggle_status
+        POST /api/v1/vendor/shop/toggle-status
         {"is_open": true}
     Example response:
         {"status": "success", "message": "Shop marked as opened"}
@@ -415,76 +414,76 @@ POST /consumer/order/return/raise/<order_id>
 
 Order (Vendor)
 --------------
-GET /order/vendor
+GET /vendor/orders
     - Auth + vendor role. Lists orders for vendor's shop.
     Example request:
-        GET /api/v1/order/vendor
+        GET /api/v1/vendor/orders
     Example response:
         {"status": "success", "orders": [...]} 
 
-POST /order/vendor/status/<order_id>
+POST /vendor/orders/<order_id>/status
     - Auth + vendor role with deliver_order permission. Body: {"status": "accepted"|"rejected"|"delivered"}. Updates order status and credits wallet on delivery if prepaid.
     Example request:
-        POST /api/v1/order/vendor/status/1
+        POST /api/v1/vendor/orders/1/status
         {"status": "delivered"}
     Example response:
         {"status": "success", "message": "Order marked as delivered"}
 
-POST /order/vendor/modify/<order_id>
+POST /vendor/orders/<order_id>/modify
     - Auth + vendor role with modify_order permission. Body: {"modifications": [{"item_id":id, "quantity":qty} ...]} Adjusts order items and sets awaiting_consumer_confirmation.
     Example request:
-        POST /api/v1/order/vendor/modify/1
+        POST /api/v1/vendor/orders/1/modify
         {"modifications": [{"item_id": 2, "quantity": 1}]}
     Example response:
         {"status": "success", "message": "Order modified"}
 
-POST /order/vendor/cancel/<order_id>
+POST /vendor/orders/<order_id>/cancel
     - Auth + vendor role with cancel_order_vendor permission. Cancels order and refunds consumer if prepaid.
     Example request:
-        POST /api/v1/order/vendor/cancel/1
+        POST /api/v1/vendor/orders/1/cancel
     Example response:
         {"status": "success", "message": "Order cancelled"}
 
-POST /order/vendor/message/send/<order_id>
+POST /vendor/orders/<order_id>/message
     - Auth + vendor role. Body: message. Sends message to consumer.
     Example request:
-        POST /api/v1/order/vendor/message/send/1
+        POST /api/v1/vendor/orders/1/message
         {"message": "Out for delivery"}
     Example response:
         {"status": "success", "message": "Message sent"}
 
-GET /order/vendor/messages/<order_id>
+GET /vendor/orders/<order_id>/messages
     - Auth + vendor role. Returns all messages for order.
     Example request:
-        GET /api/v1/order/vendor/messages/1
+        GET /api/v1/vendor/orders/1/messages
     Example response:
         {"status": "success", "messages": [...]} 
 
-GET /order/vendor/issues
-    - Auth + vendor role. Lists issues for a specific order if provided via query param order_id (optional in code).
+GET /vendor/orders/issues
+    - Auth + vendor role. Lists issues for a specific order if provided via query param order_id.
     Example request:
-        GET /api/v1/order/vendor/issues?order_id=1
+        GET /api/v1/vendor/orders/issues?order_id=1
     Example response:
         {"status": "success", "issues": [...]} 
 
-POST /return/vendor/accept/<order_id>
+POST /vendor/orders/<order_id>/return/accept
     - Auth + vendor role. Accepts pending return requests for order.
     Example request:
-        POST /api/v1/return/vendor/accept/1
+        POST /api/v1/vendor/orders/1/return/accept
     Example response:
         {"status": "success", "message": "Return request accepted"}
 
-POST /return/complete/<order_id>
+POST /vendor/orders/<order_id>/return/complete
     - Auth + vendor role. Marks accepted return as completed and issues refunds.
     Example request:
-        POST /api/v1/return/complete/1
+        POST /api/v1/vendor/orders/1/return/complete
     Example response:
         {"status": "success", "message": "Return marked as completed"}
 
-POST /return/vendor/initiate/<order_id>
+POST /vendor/orders/<order_id>/return/initiate
     - Auth + vendor role. Initiates and accepts return for specified items in delivered order.
     Example request:
-        POST /api/v1/return/vendor/initiate/1
+        POST /api/v1/vendor/orders/1/return/initiate
         {"items": [{"item_id": 1, "quantity": 1}], "reason": "expired"}
     Example response:
         {"status": "success", "message": "Return initiated and accepted"}
